Do behaviour for science labs "the right way"
Before this we were modelling science labs in the background as
basically EC -> Research Time. This works fine, and is easy, but we want
the ability for people to model more complex resource types and science
labs are a good way to test out the infrastructure for that.

As it turns out, quite a few changes were needed here to get things
working to my satisfaction:
- ConverterBehaviour now has an OnRatesComputed method that is called
  before an OnChangepoint callback. This is used by the science
  converter behaviour to compute an appropriate changepoint based on the
  rate.
- ConverterBehaviour now stores a reference to the current vessel that
  can be used by the behaviour.

This change also bundles together a bunch of other changes to get things
working properly.