Properly calculate solar flux for shadowed vessels with Kopernicus
Kopernicus sets the solar flux to 0 when the active vessel is in shadow
(I'm also not sure whether background vessels get updated, likely not).

This commit changes it so that we directly compute the solar flux for
the current position for each star and use that to determine the power
of a solar panel.