Significantly improve performance of the USI-LS status panel patches
Before this commit, there was no caching whatsoever of vessel states. We
were:
- recording the entire state of every loaded vessel, and,
- simulating each vessel until steady state.

And we were doing both of those every time the status window updated.
This was quite noticeable when warping since then the window would be
updated every frame but it is likely it would get bad pretty fast with
multiple vessels flying around.

This commit makes a few changes that should help here:
- We cache the times at which supplies and EC will be exhausted, along
  with a few other things, and then use that to interpolate. If the
  vessel hasn't had a changepoint since we last cached anything then
  we can just reuse the cache.
- We cache the simulation for the current vessel for 1 real-time second
  and use that to interpolate. This lets us reduce the amount of
  overhead for loaded vessels while still updating fast enough that the
  numbers remain accurate.