Reimplement ReadOnlyList to avoid unnecessary allocations
ReadOnlyCollection is almost exactly what we want when exposing the
inventories and converters collections from BackgroundResourceProcessor.
The one problem is that creating one requires a memory allocation every
single time it is called.

This change reimplements ReadOnlyCollection as a struct called
ReadOnlyList, which avoids the allocation and also adds inlining
annotations to its methods.