Replace System.Memory nuget reference with an ILMerged assembly
Other mods are including old versions of various dependencies of
System.Memory. These don't seem to cause errors in practice in KSP but
now that we depend on KSPBurst (one of the problematic mods) it means
that we cannot run tests or the CLI.

This adds an ILMerged version of System.Memory directly to the repo,
along with a project that builds it and uses that dll.