Delete vacuous constraints as part of the presolve
There are some constraints that are unconditionally true. It makes no
sense to pass these on to the actual simplex solver since they add no
new information.

This commit also performs strength reduction on disjunctions. If one
branch is unsolvable then they can be reduced to a constraint.
Alternatively, if one branch is guaranteed to be true then they can be
removed entirely.