Add burst-optimized computations of orbit terminators
This builds on top of the existing burst-compatible orbit methods
reimplements the orbit terminator computation as a (mostly)
straightforward newton's method solver directly working on UT.

Most of the assumptions should work even in weirder configurations that
come from kopernicus, which differs from the previous implementation.

If the orbit doesn't actually intersect the umbra then this currently
just returns a time one orbit into the future.